feat: update RationInput and RationStock models to use Float for amount and stock fields

+++ Antonieta.Backend/models/ration_stock_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, Float
 from sqlalchemy.orm import relationship
 from database import Base
 
@@ -9,6 +9,6 @@ class RationStock(Base):
     name = Column(String, unique=True, index=True)  # Nome/Tipo da ração
     description = Column(String, index=True)
     unit = Column(String, index=True)
-    stock = Column(Integer, index=True)
+    stock = Column(Float, index=True)
     distributions = relationship("Distribution", back_populates="ration")
     inputs = relationship("RationInput", back_populates="ration_stock")
